chore: Dockerfile for local Linux development
The CI build depends on the Microsoft Azure VM image for Ubuntu, which
is far from a standard environment. ARM builds also require sudo and
seem to assume that the build is being run as root in the first place.

This Dockerfile provides a fixed environment for building on Linux.